Ajustes no service

+++ src/tasks/tasks.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException, HttpCode } from '@nestjs/common';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { PrismaService } from 'src/database/prisma.service';
@@ -22,6 +22,16 @@ export class TasksService {
   }
 
   async findOne(id: string) {
+    const taskExists = await this.prisma.tasks.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!taskExists) {
+      throw new NotFoundException(`Task with ID ${id} not found`);
+    }
+
     return this.prisma.tasks.findUnique({
       where: {
         id,
@@ -30,6 +40,16 @@ export class TasksService {
   }
 
   async update(id: string, { title, description }: UpdateTaskDto) {
+    const taskExists = await this.prisma.tasks.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!taskExists) {
+      throw new NotFoundException(`Task with ID ${id} not found`);
+    }
+
     const task = await this.prisma.tasks.update({
       where: {
         id,
@@ -42,7 +62,17 @@ export class TasksService {
     return task;
   }
 
+  @HttpCode(204)
   async remove(id: string) {
+    const taskExists = await this.prisma.tasks.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!taskExists) {
+      throw new NotFoundException(`Task with ID ${id} not found`);
+    }
     await this.prisma.tasks.delete({
       where: {
         id,
